Cleaned up `included?` logic
If `included` query string item is set to `false` or not set, proceed as if the file were included on a page using `require_once()` because it was. Duh.

If the item is set to `true`, set a header and output some JSON for the obvious AJAX request.

I decided not to use the `x-http-requested-with` header just to err on the side of compatibility. So sue me. Whatever. Who cares.